wish-list-service/src/main/java/com/oshada/wishlistservice/config/JWTTokenGeneratorImpl.java
package com.oshada.wishlistservice.config;

import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;
import io.jsonwebtoken.io.Decoders;
import io.jsonwebtoken.security.Keys;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Service;

import java.security.Key;
import java.util.Date;
import java.util.HashMap;
import java.util.Map;

/*
 * This class is implementing the JWTTokenGenerator interface. This class has to be annotated with
 * @Service annotation.
 * @Service indicates annotated class is a service
 * which hold business logic in the Service layer
 *
 * */
@Service
public class JWTTokenGeneratorImpl implements JWTTokenGenerator {

    /**
     * To get the property values
     */
    @Value("${jwt.secret}")
    private String secret;

    @Value("${app.jwttoken.message}")
    private String message;

    @Override
    public Map<String, String> generateToken() {
        String jwtToken = "";
        /*
         * Generate JWT token and store in String jwtToken
         */
        Map<String, Object> claims = new HashMap<>();
//        claims.put("role", user.getRole());
        jwtToken =  Jwts
                .builder()
                .setClaims(claims)
//                .setSubject(user.getId())
                .setIssuedAt(new Date(System.currentTimeMillis()))
                .setExpiration(new Date(System.currentTimeMillis() + 1000 * 60 * 24))
                .signWith(getSigningKey(), SignatureAlgorithm.HS256)
                .compact();
        Map<String, String> jwtTokenMap = new HashMap<>();
        jwtTokenMap.put("message", "Login Successful");
        jwtTokenMap.put("token", jwtToken);

        return jwtTokenMap;
    }

    private Key getSigningKey() {
        byte[] keyBytes = Decoders.BASE64.decode(secret);
        return Keys.hmacShaKeyFor(keyBytes);
    }

}
